[step2] refactor: CustomStringParser.kt object 클래스로 변경
- pr 리뷰 반영 : https://github.com/next-step/kotlin-racingcar/pull/431#discussion_r864586106

+++ src/test/kotlin/step2/calculator/CalculatorTest.kt
@@ -26,7 +26,7 @@ class CalculatorTest {
         delimiter = '='
     )
     fun `연산 테스트`(input: String, result: String) {
-        assertThat(calculator.calculate(input)).isEqualTo(result.toLong())
+        assertThat(calculator.calculate(input)).isEqualTo(result.toDouble())
     }
 
     @Test
